feat(pingdom): add Pingdom provider backend

+++ internal/service/service.go
@@ -13,8 +13,9 @@ import (
 type UptimeCheckOption func(*UptimeCheckService) *UptimeCheckService
 
 type UptimeCheckService struct {
-	provider UptimeProvider
-	slack    *Slack
+	provider      UptimeProvider
+	slack         *Slack
+	enableDeletes bool
 }
 
 func New(options ...UptimeCheckOption) *UptimeCheckService {
@@ -55,6 +56,13 @@ func WithSlack(slackWebhookURL string, slackChannel string) UptimeCheckOption {
 	}
 }
 
+func WithDeletes(enableDeletes bool) UptimeCheckOption {
+	return func(service *UptimeCheckService) *UptimeCheckService {
+		service.enableDeletes = enableDeletes
+		return service
+	}
+}
+
 func (r *UptimeCheckService) Mutate(ctx context.Context, mutation m.Mutation, ingressName string, annotations map[string]string) {
 	check, err := m.NewUptimeCheck(ingressName, annotations)
 	if err != nil {
@@ -65,9 +73,22 @@ func (r *UptimeCheckService) Mutate(ctx context.Context, mutation m.Mutation, in
 		err = r.provider.CreateOrUpdateCheck(*check)
 		r.logMutation(ctx, err, mutation, check)
 	} else if mutation == m.Delete {
+		if !r.enableDeletes {
+			r.logDeleteDisabled(ctx, check, err)
+			return
+		}
 		err = r.provider.DeleteCheck(*check)
 		r.logMutation(ctx, err, mutation, check)
 	}
+}
+
+func (r *UptimeCheckService) logDeleteDisabled(ctx context.Context, check *m.UptimeCheck, err error) {
+	msg := fmt.Sprintf("delete of uptime check '%s' (id: %s) not executed since 'enable-deletes=false'.", check.Name, check.ID)
+	log.FromContext(ctx).Error(err, msg, "check", check)
+	if r.slack == nil {
+		return
+	}
+	r.slack.Send(ctx, ":information_source: "+msg)
 }
 
 func (r *UptimeCheckService) logAnnotationErr(ctx context.Context, err error) {
